Move Cycle component into dedicated file
